+++ mycodo/inputs/sht31_smart_gadget.py
@@ -1,14 +1,33 @@
 # coding=utf-8
+import datetime
 import logging
 import time
 
-import struct
+from flask_babel import lazy_gettext
 
+from mycodo.utils.influx import parse_measurement
+from mycodo.utils.influx import write_influxdb_value
+from mycodo.databases.models import Conversion
 from mycodo.databases.models import DeviceMeasurements
 from mycodo.inputs.base_input import AbstractInput
 from mycodo.inputs.sensorutils import calculate_dewpoint
 from mycodo.inputs.sensorutils import calculate_vapor_pressure_deficit
 from mycodo.utils.database import db_retrieve_table_daemon
+
+
+def constraints_pass_positive_value(value):
+    """
+    Check if the user input is acceptable
+    :param value: float
+    :return: tuple: (bool, list of strings)
+    """
+    errors = []
+    all_passed = True
+    # Ensure value is positive
+    if value <= 0:
+        all_passed = False
+        errors.append("Must be a positive value")
+    return all_passed, errors
 
 # Measurements
 measurements_dict = {
@@ -21,10 +40,14 @@ measurements_dict = {
         'unit': 'percent'
     },
     2: {
+        'measurement': 'battery',
+        'unit': 'percent'
+    },
+    3: {
         'measurement': 'dewpoint',
         'unit': 'C'
     },
-    3: {
+    4: {
         'measurement': 'vapor_pressure_deficit',
         'unit': 'Pa'
     }
@@ -32,15 +55,16 @@ measurements_dict = {
 
 # Input information
 INPUT_INFORMATION = {
-    'input_name_unique': 'SMART_GADGET_SHT31',
+    'input_name_unique': 'SHT31_SMART_GADGET',
     'input_manufacturer': 'Sensorion',
-    'input_name': 'Smart Gadget (SHT31)',
+    'input_name': 'SHT31 Smart Gadget',
     'measurements_name': 'Humidity/Temperature',
     'measurements_dict': measurements_dict,
 
     'options_enabled': [
         'bt_location',
         'measurements_select',
+        'custom_options',
         'period',
         'pre_output'
     ],
@@ -52,7 +76,25 @@ INPUT_INFORMATION = {
 
     'interfaces': ['BT'],
     'bt_location': '00:00:00:00:00:00',
-    'bt_adapter': '0'
+    'bt_adapter': '0',
+
+    'custom_options': [
+        {
+            'id': 'download_stored_data',
+            'type': 'bool',
+            'default_value': True,
+            'name': lazy_gettext('Download Stored Data'),
+            'phrase': lazy_gettext('Download the data logged to the device.')
+        },
+        {
+            'id': 'logging_interval',
+            'type': 'integer',
+            'default_value': 600,
+            'constraints_pass': constraints_pass_positive_value,
+            'name': lazy_gettext('Set Logging Interval'),
+            'phrase': lazy_gettext('Set the logging interval (seconds) the device will store measurements on its internal memory.')
+        }
+    ]
 }
 
 
@@ -61,27 +103,17 @@ class InputModule(AbstractInput):
     A support class for Sensorion's SHT31 Smart Gadget
 
     """
-
     def __init__(self, input_dev, testing=False):
         super(InputModule, self).__init__()
         self.logger = logging.getLogger("mycodo.inputs.sht31_smart_gadget")
+        self.unique_id = input_dev.unique_id
         self._measurements = None
-
-        self.uuid_humidity_service = '00001234-B38D-4985-720E-0F993A68EE41'
-        self.uuid_humidity_char = '00001235-B38D-4985-720E-0F993A68EE41'
-
-        self.uuid_temperature_service = '00002234-B38D-4985-720E-0F993A68EE41'
-        self.uuid_temperature_char = '00002235-B38D-4985-720E-0F993A68EE41'
-
-        self.uuid_logger_service = '0000F234-B38D-4985-720E-0F993A68EE41'
-        self.uuid_sync_time_ms = '0000F235-B38D-4985-720E-0F993A68EE41'
-        self.uuid_oldest_timestamp_ms = '0000F236-B38D-4985-720E-0F993A68EE41'
-        self.uuid_newest_timestamp_ms = '0000F237-B38D-4985-720E-0F993A68EE41'
-        self.uuid_start_logger_download = '0000F238-B38D-4985-720E-0F993A68EE41'
-        self.uuid_logger_interval_ms_char = '0000F239-B38D-4985-720E-0F993A68EE41'
+        self.download_stored_data = None
+        self.logging_interval = None
+        self.gadget = None
 
         if not testing:
-            from bluepy.btle import Peripheral
+            from mycodo.devices.sht31_smart_gadget import SHT31
             from bluepy import btle
             self.logger = logging.getLogger(
                 "mycodo.sht31_smart_gadget_{id}".format(id=input_dev.unique_id.split('-')[0]))
@@ -90,83 +122,123 @@ class InputModule(AbstractInput):
                 DeviceMeasurements).filter(
                     DeviceMeasurements.device_id == input_dev.unique_id)
 
+            if input_dev.custom_options:
+                for each_option in input_dev.custom_options.split(';'):
+                    option = each_option.split(',')[0]
+                    value = each_option.split(',')[1]
+                    if option == 'download_stored_data':
+                        self.download_stored_data = bool(value)
+                    elif option == 'logging_interval':
+                        self.logging_interval = int(value)
+
+            self.SHT31 = SHT31
             self.btle = btle
             self.location = input_dev.location
             self.bt_adapter = input_dev.bt_adapter
-
-            # Create Peripheral but do not connect yet
-            self.p = Peripheral(None,
-                                "random",
-                                iface=self.bt_adapter)
 
     def get_measurement(self):
         """ Obtain and return the measurements """
         return_dict = measurements_dict.copy()
 
-        try:
-            # Make three attempts to connect
-            connected = False
-            connect_error = None
-            for _ in range(3):
-                try:
-                    self.p.connect(
-                        self.location, "random", iface=self.bt_adapter)
-                    connected = True
-                    break
-                except self.btle.BTLEException as e:
-                    connect_error = e
-                time.sleep(0.5)
+        # Make three attempts to connect
+        connected = False
+        connect_error = None
+        for _ in range(3):
+            try:
+                self.gadget = self.SHT31(addr=self.location,
+                                         iface=self.bt_adapter)
+                connected = True
+                break
+            except self.btle.BTLEException as e:
+                connect_error = e
+            time.sleep(0.5)
 
-            if not connected:
-                self.logger.error(
-                    "Could not connect (made 3 attempts). Error: {}".format(
-                        connect_error))
-                return
-
-            sht31_service_hum = self.p.getServiceByUUID(
-                self.uuid_humidity_service)
-            ch = sht31_service_hum.getCharacteristics(
-                self.uuid_humidity_char)[0]
-            humidity = struct.unpack('<f', ch.read())[0]
-
-            sht31_service_temp = self.p.getServiceByUUID(
-                self.uuid_temperature_service)
-            ch = sht31_service_temp.getCharacteristics(
-                self.uuid_temperature_char)[0]
-            temperature = struct.unpack('<f', ch.read())[0]
-
+        if connected:
             if self.is_enabled(0):
-                return_dict[0]['value'] = temperature
+                return_dict[0]['value'] = self.gadget.readTemperature()
 
             if self.is_enabled(1):
-                return_dict[1]['value'] = humidity
+                return_dict[1]['value'] = self.gadget.readHumidity()
 
-            if (self.is_enabled(2) and
-                    self.is_enabled(0) and
-                    self.is_enabled(1)):
-                return_dict[2]['value'] = calculate_dewpoint(
-                    return_dict[0]['value'], return_dict[1]['value'])
+            if self.is_enabled(2):
+                return_dict[2]['value'] = self.gadget.readBattery()
 
             if (self.is_enabled(3) and
                     self.is_enabled(0) and
                     self.is_enabled(1)):
-                return_dict[3]['value'] = calculate_vapor_pressure_deficit(
+                return_dict[3]['value'] = calculate_dewpoint(
                     return_dict[0]['value'], return_dict[1]['value'])
 
-            try:
-                self.p.disconnect()
-            except self.btle.BTLEException as e:
-                self.logger.error("Disconnect Error: {}".format(e))
+            if (self.is_enabled(4) and
+                    self.is_enabled(0) and
+                    self.is_enabled(1)):
+                return_dict[4]['value'] = calculate_vapor_pressure_deficit(
+                    return_dict[0]['value'], return_dict[1]['value'])
 
-        except self.btle.BTLEException as e:
-            self.logger.error("Error: {}".format(e))
+            if self.download_stored_data:
 
-        return return_dict
+                # Download stored data from the device
+                self.gadget.readLoggedDataInterval()
+                while True:
+                    if (not self.gadget.waitForNotifications(5) or
+                            not self.gadget.isLogReadoutInProgress()):
+                        break  # Done reading data
 
-    def stop_sensor(self):
-        """ Called when the sensor is deactivated """
-        try:
-            self.p.disconnect()
-        except Exception as msg:
-            self.logger.error("Error: {err}".format(err=msg))
-        self.running = False
+                # Store logged temperatures
+                measurement = self.device_measurements.filter(DeviceMeasurements.channel == 0).first()
+                conversion = db_retrieve_table_daemon(Conversion, unique_id=measurement.conversion_id)
+                for each_time, each_measure in self.gadget.loggedData['Temp'].items():
+                    measurement_single = {
+                        0: {
+                            'measurement': 'temperature',
+                            'unit': 'C',
+                            'value': each_measure
+                        }
+                    }
+                    measurement_single = parse_measurement(
+                        conversion,
+                        measurement,
+                        measurement_single,
+                        measurement.channel,
+                        measurement_single[0])
+                    write_influxdb_value(
+                        self.unique_id,
+                        measurement_single[0]['unit'],
+                        value=measurement_single[0]['value'],
+                        measure=measurement_single[0]['measurement'],
+                        channel=0,
+                        timestamp=datetime.datetime.utcfromtimestamp(each_time / 1000))
+
+                # Store logged humidities
+                measurement = self.device_measurements.filter(DeviceMeasurements.channel == 1).first()
+                conversion = db_retrieve_table_daemon(Conversion, unique_id=measurement.conversion_id)
+                for each_time, each_measure in self.gadget.loggedData['Humi'].items():
+                    measurement_single = {
+                        1: {
+                            'measurement': 'humidity',
+                            'unit': '%',
+                            'value': each_measure
+                        }
+                    }
+                    measurement_single = parse_measurement(
+                        conversion,
+                        measurement,
+                        measurement_single,
+                        measurement.channel,
+                        measurement_single[1])
+                    write_influxdb_value(
+                        self.unique_id,
+                        measurement_single[1]['unit'],
+                        value=measurement_single[1]['value'],
+                        measure=measurement_single[1]['measurement'],
+                        channel=0,
+                        timestamp=datetime.datetime.utcfromtimestamp(each_time / 1000))
+
+                # Reset log and set logging interval
+                self.gadget.setLoggerIntervalMs(self.logging_interval * 1000)
+
+            self.gadget.disconnect()
+
+            return return_dict
+        else:
+            self.logger.error("Could not connect: {}".format(connect_error))