+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Features
 
  - Add Additional check for Conditional Statements if {ID} is replaced with None ([#571](https://github.com/kizniche/mycodo/issues/571))
+ - Add ability to set Logging Interval and download logged data from SHT31 Smart Gadget ([#559](https://github.com/kizniche/mycodo/issues/559))
 
 
 ## 7.0.1 (2018-12-09)